+++ src/aeolus/sources/regulatory.py
@@ -89,8 +89,10 @@ from ..transforms import (
     melt_measurands,
     rename_columns,
     reset_index,
+    select_columns,
 )
 from ..types import (
+    DATA_COLUMNS,
     AeolusDataWarning,
     DataFetcher,
     MetadataFetcher,
@@ -341,6 +343,7 @@ def normalise_regulatory_data(network_name: str) -> Normaliser:
             ),
             add_column("created_at", lambda df: datetime.now(timezone.utc)),
             drop_columns("site_name"),
+            select_columns(*DATA_COLUMNS, require_all=True),
         )(df)
 
     return normalise